refactor / add sign in / sign out and home to the navbar items. removed googleSignIn and googleSignOut components, in favour of a useGoogleAuth hook
